+++ src/codegen/utilities/interpreter_helpers.py
@@ -68,6 +68,7 @@ GRPC_INTERPRETER_IGNORED_FUNCTIONS = [
     "get_error_string",
     "read_id_pin_memory",
     "set_runtime_environment",
+    "internal_get_last_created_chan",
 ]
 
 LIBRARY_INTERPRETER_IGNORED_FUNCTIONS = [